Use caml_enter_blocking_section() for some stubs
This patch caml_enter_blocking_section() and
caml_leave_blocking_section() to the following C stub functions so that
they do not block the domain that they are running in:

* ml_al_wait_for_event
* ml_al_wait_for_event_timed
* ml_al_wait_for_event_until
* ml_al_wait_for_event_rest

This also adds some small test cases for this functionality. The tests
are disabled by default but can be enabled by setting the RUN_GUI_TEST
environment variable to true.

The test scenario for all of these is similar: launch a thread
that delays 1 second and then updates a ref variable. Call one of the
modified functions and have it block the thread for 3 seconds. Verify
that the first thread executed during that time by checking the ref
variable.

* test_al_rest.ml
* test_al_wait_for_event_until.ml
* test_al_wait_for_event_timed.ml
* test_al_wait_for_event.ml